Add end points for global causal policy and global causal effects
