+++ raiwidgets/raiwidgets/responsibleai_dashboard.py
@@ -5,10 +5,10 @@
 
 from flask import jsonify, request
 
+from raiwidgets.dashboard import Dashboard
+from raiwidgets.responsibleai_dashboard_input import \
+    ResponsibleAIDashboardInput
 from responsibleai import RAIInsights
-
-from .dashboard import Dashboard
-from .responsibleai_dashboard_input import ResponsibleAIDashboardInput
 
 
 class ResponsibleAIDashboard(Dashboard):
@@ -66,11 +66,18 @@ class ResponsibleAIDashboard(Dashboard):
 
         self.add_url_rule(causal_whatif, '/causal_whatif', methods=["POST"])
 
-        def causal_global_effects():
+        def global_causal_effects():
             data = request.get_json(force=True)
-            return jsonify(self.input.causal_global_effects(data))
+            return jsonify(self.input.get_global_causal_effects(data))
 
-        self.add_url_rule(causal_global_effects, '/causal_global_effects',
+        self.add_url_rule(global_causal_effects, '/global_causal_effects',
+                          methods=["POST"])
+
+        def global_causal_policy():
+            data = request.get_json(force=True)
+            return jsonify(self.input.get_global_causal_policy(data))
+
+        self.add_url_rule(global_causal_policy, '/global_causal_policy',
                           methods=["POST"])
 
         def importances():
