+++ raiwidgets/tests/responsibleai_dashboard/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) Microsoft Corporation
+# Licensed under the MIT License.
+
+"""RAIWidgets ResponsibleAIDashboard tests."""
